refactor(#183): migrate runtime name policy module to runtime path
