+++ sdk/src/query/helpers.ts
@@ -24,7 +24,7 @@ import { homedir } from 'node:os';
 import { GSDError, ErrorClassification } from '../errors.js';
 export { SUPPORTED_RUNTIMES, type Runtime } from '../model-catalog.js';
 import { SUPPORTED_RUNTIMES, type Runtime } from '../model-catalog.js';
-import { canonicalizeRuntimeName } from '../runtime-name-policy.js';
+import { canonicalizeRuntimeName } from '../runtime/name-policy.js';
 import { workspacePlanningPaths, resolveWorkspaceContext, type PlanningPaths } from './workspace.js';
 export { stateExtractField } from '../state/index.js';
 import { relPlanningPath, validateWorkstreamName } from '../workstream-utils.js';
